+++ FireProductManager/ServiceLogicPackage/马华/BarrelGateway.cs
@@ -36,7 +36,7 @@ namespace FireProductManager.ServiceLogicPackage
             Barrel barrel = new Barrel();//id自加
             barrel.ba_isRemoved = 0; //0为存在，1为不存在
             barrel.Insert();
-            barrel.ba_id = (uint)FindMaxBarrelId();
+            barrel.ba_id = FindMaxBarrelId();
             return (int)barrel.ba_id;
         }
 
@@ -46,7 +46,7 @@ namespace FireProductManager.ServiceLogicPackage
             if (!IsBarrelIdValid(barrelid)) return false;
 
             Barrel barrel = new Barrel();
-            barrel.ba_id = (uint)barrelid;
+            barrel.ba_id = barrelid;
             barrel.ba_isRemoved = 1;
             barrel.Update();
             return true;
